software-engineering/atproto/_src/python-atproto/pyproject.toml
[project]
name = "python-atproto"
version = "0.1.0"
description = "Add a short description here"
authors = [
    { name = "xhiroga", email = "xhiroga@users.noreply.github.com" }
]
dependencies = [
    "atproto~=0.0.16",
    "python-dotenv~=1.0.0",
    "openai~=0.27.8",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true

[tool.hatch.metadata]
allow-direct-references = true

[tool.rye.scripts]
getting_started = { cmd = "python ./src/python_atproto/getting_started.py" }
firehose = { cmd = "python ./src/python_atproto/firehose.py" }
home_timeline = { cmd = "python ./src/python_atproto/home_timeline.py" }
